Socket handle structure 2
Extra package for both client and server. It is required to be installed through package manager. Already in requirements.txt files

Socket Thread Factory server class that identifies sockets from client and chooses correct socket thread for server

Authorization sockets for both client and server. For signing up and signing in

Minimalistic Database module. It can compare login and password with db and save login and password if login is not in db

ABCs for Threads, including Socket Threads

Some docs